refactor: move MB to shared/constants and use MB throughout app
