+++ src/app/models/form.server.model.ts
@@ -9,6 +9,7 @@ import mongoose, {
 } from 'mongoose'
 import validator from 'validator'
 
+import { MB } from '../../shared/constants'
 import { reorder } from '../../shared/util/immutable-array-fns'
 import {
   AuthType,
@@ -41,7 +42,6 @@ import {
   Status,
 } from '../../types'
 import { IPopulatedUser, IUserSchema } from '../../types/user'
-import { MB } from '../constants/filesize'
 import { OverrideProps } from '../modules/form/admin-form/admin-form.types'
 import { getFormFieldById, transformEmails } from '../modules/form/form.utils'
 import { validateWebhookUrl } from '../modules/webhook/webhook.validation'
