+++ src/app/modules/submission/email-submission/email-submission.receiver.ts
@@ -2,9 +2,9 @@ import Busboy from 'busboy'
 import { IncomingHttpHeaders } from 'http'
 import { err, ok, Result, ResultAsync } from 'neverthrow'
 
+import { MB } from '../../../../shared/constants'
 import { IAttachmentInfo } from '../../../../types'
 import { createLoggerWithLabel } from '../../../config/logger'
-import { MB } from '../../../constants/filesize'
 
 import {
   InitialiseMultipartReceiverError,
