+++ src/app/utils/field-validation/validators/attachmentValidator.ts
@@ -2,10 +2,9 @@ import { chain, left, right } from 'fp-ts/lib/Either'
 import { flow } from 'fp-ts/lib/function'
 
 import { ProcessedAttachmentResponse } from 'src/app/modules/submission/submission.types'
+import { MB } from 'src/shared/constants'
 import { IAttachmentField } from 'src/types/field'
 import { ResponseValidator } from 'src/types/field/utils/validation'
-
-const MILLION = 1000000
 
 type AttachmentValidator = ResponseValidator<ProcessedAttachmentResponse>
 type AttachmentValidatorConstructor = (
@@ -38,14 +37,15 @@ const attachmentContentValidator: AttachmentValidator = (response) => {
  * Returns a validation function to check if
  * attachment size is within the specified limit.
  */
-const makeAttachmentSizeValidator: AttachmentValidatorConstructor =
-  (attachmentField) => (response) => {
-    const { attachmentSize } = attachmentField
-    const byteSizeLimit = parseInt(attachmentSize) * MILLION
-    return response.content.byteLength < byteSizeLimit
-      ? right(response)
-      : left(`AttachmentValidator:\t File size more than limit`)
-  }
+const makeAttachmentSizeValidator: AttachmentValidatorConstructor = (
+  attachmentField,
+) => (response) => {
+  const { attachmentSize } = attachmentField
+  const byteSizeLimit = parseInt(attachmentSize) * MB
+  return response.content.byteLength < byteSizeLimit
+    ? right(response)
+    : left(`AttachmentValidator:\t File size more than limit`)
+}
 
 /**
  * Returns a validation function for an attachment field when called.
