+++ src/public/modules/forms/admin/controllers/edit-fields-modal.client.controller.js
@@ -7,6 +7,7 @@ const cloneDeep = require('lodash/cloneDeep')
 const {
   VALID_UPLOAD_FILE_TYPES,
   MAX_UPLOAD_FILE_SIZE,
+  MB,
 } = require('shared/constants')
 const { UPDATE_FORM_TYPES } = require('../constants/update-form-types')
 const { uploadImage } = require('../../../../services/FileHandlerService')
@@ -559,8 +560,8 @@ function EditFieldsModalController(
       field.uploadedFile = ''
       switch (ngfError.$error) {
         case 'maxSize':
-          vm.uploadError = `${(ngfError.size / 1000000).toFixed(2)} MB / ${
-            vm.maxImageSize / 1000000
+          vm.uploadError = `${(ngfError.size / MB).toFixed(2)} MB / ${
+            vm.maxImageSize / MB
           } MB: File size exceeded`
           break
         case 'resize':
@@ -585,7 +586,7 @@ function EditFieldsModalController(
           field.url = result.url
           field.fileMd5Hash = result.fileMd5Hash
           field.name = result.name
-          field.size = `${(result.size / 1000000).toFixed(2)} MB`
+          field.size = `${(result.size / MB).toFixed(2)} MB`
         })
         .catch((uploadError) => {
           // This is a reference to the ng-model of the upload button, which points to the uploaded file
