+++ src/public/modules/forms/admin/views/edit-start-page.client.modal.html
@@ -79,7 +79,7 @@
                   >
                   <span id="image-size"
                     >{{
-                    (vm.myform.startPage.logo.fileSizeInBytes/1000000).toFixed(2)
+                    (vm.myform.startPage.logo.fileSizeInBytes/vm.MB).toFixed(2)
                     }} MB</span
                   >
                 </div>
