+++ src/public/modules/forms/base/components/field-attachment.client.component.js
@@ -5,7 +5,7 @@ const {
   isInvalidFileExtension,
   getInvalidFileExtensionsInZip,
 } = require('../../../../../shared/util/file-validation')
-const { FilePlatforms } = require('../../../../../shared/constants')
+const { FilePlatforms, KB, MB } = require('../../../../../shared/constants')
 
 angular.module('forms').component('attachmentFieldComponent', {
   templateUrl:
@@ -22,6 +22,8 @@ angular.module('forms').component('attachmentFieldComponent', {
 function attachmentFieldComponentController($timeout) {
   const vm = this
 
+  vm.MB = MB
+
   vm.isAttachmentTouched = false
   vm.touchAttachment = function () {
     vm.isAttachmentTouched = true
@@ -37,7 +39,7 @@ function attachmentFieldComponentController($timeout) {
     if (errFiles.length > 0) {
       err = errFiles[0].$error
       if (err === 'maxSize') {
-        const currentSize = (errFiles[0].size / 1000000).toFixed(2)
+        const currentSize = (errFiles[0].size / MB).toFixed(2)
         showAttachmentError(
           `${currentSize} MB / ${vm.field.attachmentSize} MB: File size exceeded`,
         )
@@ -136,11 +138,11 @@ function attachmentFieldComponentController($timeout) {
         vm.fileAttached = true
         vm.fileError = false
         vm.fileName = file.name
-        vm.fileSize = file.size / 1000
-        if (file.size / 1000 > 1000) {
-          vm.fileSize = String((file.size / 1000000).toFixed(2)) + ' MB'
+        vm.fileSize = file.size / KB
+        if (file.size / KB > KB) {
+          vm.fileSize = String((file.size / MB).toFixed(2)) + ' MB'
         } else {
-          vm.fileSize = String((file.size / 1000).toFixed(2)) + ' KB'
+          vm.fileSize = String((file.size / KB).toFixed(2)) + ' KB'
         }
         vm.isLoading = false
       })
