+++ src/shared/constants.ts
@@ -22,8 +22,11 @@ export enum DateSelectedValidation {
 // File types that can be uploaded for form image/logo
 export const VALID_UPLOAD_FILE_TYPES = ['image/jpeg', 'image/png', 'image/gif']
 
+export const KB = 1000
+export const MB = 1000 * KB
+
 // Define max file size as 2MB
-export const MAX_UPLOAD_FILE_SIZE = 2 * 1000 * 1000 // 2 Million/Mega Bytes, or 2 MB
+export const MAX_UPLOAD_FILE_SIZE = 2 * MB // 2 Million/Mega Bytes, or 2 MB
 
 export const LINKS = {
   supportFormLink: 'https://go.gov.sg/formsg-support',
